refactor(perks,kyc,request): decompose PerkClaimModal, SumsubKycWrapper and Create.request.link.view in place (TASK-21854)
each monolith splits into a flow/lifecycle hook plus named subview
components in sibling files; composition roots keep their paths and
exports so importers are untouched. the request-link url params move to
nuqs (new-file ratchet). zero behavior change otherwise.
