+++ src/components/Home/PerkClaimModal.tsx
@@ -1,30 +1,10 @@
 'use client'
 
-import { useState, useCallback, useRef, useEffect } from 'react'
-import { useTranslations } from 'next-intl'
-import { useAppTranslations } from '@/i18n/app/useAppTranslations'
-import { useQueryClient } from '@tanstack/react-query'
-import { perksApi, type PendingPerk } from '@/services/perks'
-import { Icon } from '@/components/Global/Icons/Icon'
-import { useHoldToClaim } from '@/hooks/useHoldToClaim'
-import { getShakeClass } from '@/utils/perk.utils'
-import { extractInviteeName } from '@/utils/general.utils'
-import { shootDoubleStarConfetti } from '@/utils/confetti'
-import { notifyHaptic } from '@/utils/haptics'
-import { SoundPlayer } from '@/components/Global/SoundPlayer'
-import { useAppHaptic } from '@/hooks/useAppHaptic'
+import { type PendingPerk } from '@/services/perks'
 import ActionModal from '@/components/Global/ActionModal'
-import { Button } from '@/components/0_Bruddle/Button'
-import InviteFriendsModal from '@/components/Global/InviteFriendsModal'
-import { useRouter } from 'next/navigation'
-import { getUserPreferences, updateUserPreferences } from '@/utils/general.utils'
-import { useAuth } from '@/context/authContext'
-import posthog from 'posthog-js'
-import { ANALYTICS_EVENTS, REFERRAL_SOURCES } from '@/constants/analytics.consts'
-
-type ClaimPhase = 'idle' | 'holding' | 'opening' | 'revealed' | 'exiting'
-
-const SURPRISE_CLAIM_COUNT_KEY = 'rewards_surprise_claim_count'
+import { usePerkClaimFlow } from './usePerkClaimFlow'
+import { PerkClaimSuccessModal } from './PerkClaimSuccessModal'
+import { PerkClaimGiftBox } from './PerkClaimGiftBox'
 
 interface PerkClaimModalProps {
     perk: PendingPerk
@@ -39,112 +19,15 @@ interface PerkClaimModalProps {
  * Uses ActionModal for consistent styling with other modals.
  */
 function PerkClaimModal({ perk, visible, onClose, onClaimed }: PerkClaimModalProps) {
-    const queryClient = useQueryClient()
-    const { user } = useAuth()
-    const [claimPhase, setClaimPhase] = useState<ClaimPhase>('idle')
-    const [lastClaimedPerk, setLastClaimedPerk] = useState<PendingPerk | null>(null)
-    const apiCallRef = useRef<Promise<void> | null>(null)
-    const revealTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-    const dismissTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-
-    // Cleanup timers on unmount
-    useEffect(() => {
-        return () => {
-            if (revealTimerRef.current) clearTimeout(revealTimerRef.current)
-            if (dismissTimerRef.current) clearTimeout(dismissTimerRef.current)
-        }
-    }, [])
-
-    // Reset state when modal opens with new perk
-    useEffect(() => {
-        if (visible) {
-            setClaimPhase('idle')
-            setLastClaimedPerk(null)
-
-            const eventProps = { amount_usd: perk.amountUsd, perk_name: perk.name }
-            posthog.capture(ANALYTICS_EVENTS.REWARD_CLAIM_SHOWN, eventProps)
-
-            // Read claim count from user prefs to determine if this is a surprise moment
-            const userId = user?.user.userId ?? ''
-            const claimCount = getUserPreferences(userId)?.[SURPRISE_CLAIM_COUNT_KEY] ?? 0
-            if (claimCount < 2) {
-                posthog.capture(ANALYTICS_EVENTS.SURPRISE_MOMENT_SHOWN, {
-                    ...eventProps,
-                    claim_number: claimCount + 1,
-                })
-            }
-        }
-        // perk props and user are stable while modal is open — deps kept minimal to fire once per open
-    }, [visible, perk.id, perk.amountUsd, perk.name, user?.user.userId])
-
-    // Optimistic claim: trigger animation immediately, API call in background
-    const handleHoldComplete = useCallback(async () => {
-        // Phase 1: Opening animation (gift shakes on its own, builds anticipation)
-        setClaimPhase('opening')
-
-        // Fire API call in background - don't await it
-        apiCallRef.current = (async () => {
-            try {
-                const result = await perksApi.claimPerk(perk.id)
-                if (result.success) {
-                    posthog.capture(ANALYTICS_EVENTS.REWARD_CLAIMED, {
-                        amount_usd: perk.amountUsd,
-                        perk_name: perk.name,
-                        invitee_name: perk.inviteeName ?? extractInviteeName(perk.reason),
-                    })
-                    onClaimed(perk.id)
-                    queryClient.invalidateQueries({ queryKey: ['pendingPerks'] })
-                    queryClient.invalidateQueries({ queryKey: ['transactions'] })
-                }
-            } catch (error) {
-                console.error('Failed to claim perk:', error)
-            }
-        })()
-
-        // Phase 2: After 600ms of autonomous shaking, burst into confetti
-        revealTimerRef.current = setTimeout(() => {
-            // Haptic burst feedback
-            notifyHaptic('success')
-
-            // Confetti explosion!
-            shootDoubleStarConfetti({ origin: { x: 0.5, y: 0.4 } })
-
-            // Phase 3: Show revealed state
-            setLastClaimedPerk(perk)
-            setClaimPhase('revealed')
-        }, 600)
-    }, [perk, queryClient, onClaimed])
-
-    // Handle dismissing the success message
-    const handleDismissSuccess = useCallback(() => {
-        setClaimPhase('exiting')
-        dismissTimerRef.current = setTimeout(() => {
-            onClose()
-        }, 400)
-    }, [onClose])
-
-    // Handle modal close based on current phase
-    const handleModalClose = useCallback(() => {
-        if (claimPhase === 'revealed') {
-            handleDismissSuccess()
-        } else if (claimPhase === 'idle') {
-            posthog.capture(ANALYTICS_EVENTS.REWARD_CLAIM_DISMISSED, {
-                amount_usd: perk.amountUsd,
-                perk_name: perk.name,
-            })
-            onClose()
-        }
-        // Don't allow closing during opening/exiting phases
-    }, [claimPhase, handleDismissSuccess, onClose, perk])
+    const { claimPhase, lastClaimedPerk, isSuccessPhase, handleHoldComplete, handleDismissSuccess, handleModalClose } =
+        usePerkClaimFlow({ perk, visible, onClose, onClaimed })
 
     if (!visible) return null
-
-    const isSuccessPhase = (claimPhase === 'revealed' || claimPhase === 'exiting') && !!lastClaimedPerk
 
     // Use ActionModal's native props for success phase, custom content for gift box phase
     if (isSuccessPhase) {
         return (
-            <SuccessModal
+            <PerkClaimSuccessModal
                 perk={lastClaimedPerk!}
                 claimPhase={claimPhase}
                 onClose={handleModalClose}
@@ -159,307 +42,8 @@ function PerkClaimModal({ perk, visible, onClose, onClaimed }: PerkClaimModalPro
             onClose={handleModalClose}
             title=""
             preventClose={claimPhase === 'opening'}
-            content={<GiftBoxContent perk={perk} onHoldComplete={handleHoldComplete} claimPhase={claimPhase} />}
+            content={<PerkClaimGiftBox perk={perk} onHoldComplete={handleHoldComplete} claimPhase={claimPhase} />}
         />
-    )
-}
-
-interface SuccessModalProps {
-    perk: PendingPerk
-    claimPhase: ClaimPhase
-    onClose: () => void
-    onDismiss: () => void
-}
-
-/**
- * Success modal using ActionModal's native layout for consistent design system styling.
- * Uses icon/title/description props for standard vertical centered layout.
- */
-function SuccessModal({ perk, claimPhase, onClose, onDismiss }: SuccessModalProps) {
-    const t = useAppTranslations('home.perk')
-    const tCommon = useTranslations('common')
-    const inviteeName = perk.inviteeName ?? extractInviteeName(perk.reason)
-    const { triggerHaptic } = useAppHaptic()
-    const router = useRouter()
-    const { user } = useAuth()
-    const [canDismiss, setCanDismiss] = useState(false)
-    const [isInviteModalOpen, setIsInviteModalOpen] = useState(false)
-    const isExiting = claimPhase === 'exiting'
-
-    // Surprise moment claim count: read synchronously so first render has correct copy.
-    // 0=first surprise, 1=second, 2+=normal referral claim.
-    // NOTE: stored in localStorage per-user. Cross-device users may see surprise copy again.
-    // TODO: move to BE (perk_usage count for surprise campaigns) for authoritative tracking.
-    const [claimCount] = useState(() => {
-        const prefs = user?.user.userId ? getUserPreferences(user.user.userId) : null
-        return prefs?.[SURPRISE_CLAIM_COUNT_KEY] ?? 0
-    })
-    useEffect(() => {
-        if (!user?.user.userId) return
-        updateUserPreferences(user.user.userId, { [SURPRISE_CLAIM_COUNT_KEY]: claimCount + 1 })
-    }, []) // eslint-disable-line react-hooks/exhaustive-deps -- fire once on mount
-
-    useEffect(() => {
-        triggerHaptic()
-        const dismissTimer = setTimeout(() => setCanDismiss(true), 2000)
-        return () => clearTimeout(dismissTimer)
-    }, []) // eslint-disable-line react-hooks/exhaustive-deps -- triggerHaptic is stable
-
-    const isSurpriseMoment = claimCount < 2
-
-    return (
-        <>
-            <ActionModal
-                visible={true}
-                onClose={onClose}
-                hideModalCloseButton
-                preventClose={isExiting}
-                icon="check"
-                iconProps={{ className: 'text-white' }}
-                iconContainerClassName="bg-background-icon-bubble-green"
-                title=""
-                description={
-                    <div className={isExiting ? 'animate-gift-exit' : 'animate-gift-revealed'}>
-                        <p className="text-heading-m text-black">+${perk.amountUsd}</p>
-                        {isSurpriseMoment ? (
-                            <>
-                                {/* Approved copy — see notion: notifs-copy-33083811757980638a27effc79a033f3 */}
-                                <p className="mt-2 text-center text-body-m-semibold text-foreground-primary">
-                                    {t('surpriseTitle', { amount: perk.amountUsd })}
-                                </p>
-                                <p className="mt-1 text-center text-body-s text-foreground-secondary">
-                                    {claimCount === 0 ? t('surpriseDescriptionFirst') : t('surpriseDescriptionNext')}
-                                </p>
-                            </>
-                        ) : inviteeName ? (
-                            <p className="mt-1 flex items-center justify-center gap-1 text-body-s text-foreground-secondary">
-                                <Icon name="invite-heart" size={16} />
-                                {t.rich('usedPeanut', {
-                                    inviteeName,
-                                    name: (chunks) => <span className="font-medium">{chunks}</span>,
-                                })}
-                            </p>
-                        ) : (
-                            <p className="mt-1 text-body-s text-foreground-secondary">{t('rewardClaimed')}</p>
-                        )}
-                    </div>
-                }
-                content={
-                    <>
-                        <SoundPlayer sound="success" />
-                        {canDismiss && (
-                            <div className="mt-4 flex flex-col items-center gap-2">
-                                {isSurpriseMoment ? (
-                                    <>
-                                        <Button
-                                            variant="purple"
-                                            shadowSize="4"
-                                            className="w-full"
-                                            onClick={() => {
-                                                posthog.capture(ANALYTICS_EVENTS.REFERRAL_CTA_CLICKED, {
-                                                    source: REFERRAL_SOURCES.SURPRISE_MOMENT,
-                                                })
-                                                onDismiss()
-                                                setIsInviteModalOpen(true)
-                                            }}
-                                        >
-                                            {t('shareAndEarn')}
-                                        </Button>
-                                        <Button variant="stroke" className="w-full" onClick={onDismiss}>
-                                            {tCommon('maybeLater')}
-                                        </Button>
-                                    </>
-                                ) : (
-                                    <>
-                                        <Button variant="purple" shadowSize="4" className="w-full" onClick={onDismiss}>
-                                            {tCommon('done')}
-                                        </Button>
-                                        <Button
-                                            variant="stroke"
-                                            className="w-full"
-                                            onClick={() => {
-                                                onDismiss()
-                                                router.push('/rewards')
-                                            }}
-                                        >
-                                            {t('inviteFriendsToEarnMore')}
-                                        </Button>
-                                    </>
-                                )}
-                            </div>
-                        )}
-                    </>
-                }
-            />
-            {user?.user.username && (
-                <InviteFriendsModal
-                    visible={isInviteModalOpen}
-                    onClose={() => setIsInviteModalOpen(false)}
-                    username={user.user.username}
-                    source="surprise_moment"
-                />
-            )}
-        </>
-    )
-}
-
-interface GiftBoxContentProps {
-    perk: PendingPerk
-    onHoldComplete: () => void
-    claimPhase: ClaimPhase
-}
-
-/**
- * Gift box with hold-to-claim interaction
- */
-function GiftBoxContent({ perk, onHoldComplete, claimPhase }: GiftBoxContentProps) {
-    const t = useAppTranslations('home.perk')
-    const { holdProgress, isShaking, shakeIntensity, buttonProps } = useHoldToClaim({
-        onComplete: onHoldComplete,
-        disabled: claimPhase !== 'idle',
-        enableTapMode: true,
-        tapProgress: 12,
-        holdProgressPerSec: 80,
-        decayRate: 8,
-    })
-
-    // Ribbon opens outward based on hold progress (max 30deg spread)
-    const ribbonSpread = (holdProgress / 100) * 30
-
-    // Determine animation classes based on phase
-    const getAnimationClass = () => {
-        if (claimPhase === 'opening') {
-            return 'animate-gift-opening'
-        }
-        if (isShaking) {
-            return getShakeClass(isShaking, shakeIntensity)
-        }
-        return ''
-    }
-
-    const inviteeName = perk.inviteeName ?? extractInviteeName(perk.reason)
-
-    return (
-        <div className="flex flex-col items-center">
-            {/* Title */}
-            <p className="mb-6 text-center text-body-s text-foreground-secondary">
-                <Icon name="invite-heart" size={16} className="mr-1 inline" />
-                {t.rich('usedPeanut', {
-                    inviteeName: inviteeName ?? '',
-                    name: (chunks) => <span className="font-medium">{chunks}</span>,
-                })}
-            </p>
-
-            {/* Gift box wrapper - only this shakes */}
-            <div className={`relative ${getAnimationClass()}`}>
-                {/* Glow effect behind gift */}
-                <div
-                    className="pointer-events-none absolute inset-0 -m-6 rounded-3xl bg-action-primary blur-2xl transition-opacity"
-                    style={{ opacity: (holdProgress / 100) * 0.3 }}
-                />
-
-                {/* Gift box container */}
-                <div {...buttonProps} className="relative cursor-pointer touch-none select-none">
-                    {/* Gift box */}
-                    <div
-                        className={`gift-box-shine relative h-32 w-44 overflow-hidden rounded-xl border-4 border-action-primary bg-gradient-to-br from-action-primary/20 via-white to-action-primary/20 shadow-xl transition-transform ${holdProgress > 0 ? 'scale-[0.98]' : ''}`}
-                    >
-                        {/* Vertical ribbon */}
-                        <div className="absolute top-0 bottom-0 left-1/2 w-5 -translate-x-1/2 bg-gradient-to-r from-action-primary/50 via-action-primary/70 to-action-primary/50" />
-
-                        {/* Horizontal ribbon */}
-                        <div className="absolute top-1/2 right-0 left-0 h-5 -translate-y-1/2 bg-gradient-to-b from-action-primary/50 via-action-primary/70 to-action-primary/50" />
-
-                        {/* Light rays from center */}
-                        <div
-                            className="pointer-events-none absolute inset-0"
-                            style={{
-                                background: `radial-gradient(circle at center, rgba(255,255,255,${0.4 * (holdProgress / 100)}) 0%, transparent 70%)`,
-                            }}
-                        />
-
-                        {/* Cracks appearing with progress */}
-                        {holdProgress > 20 && (
-                            <div className="absolute top-4 left-4 h-8 w-0.5 rotate-45 bg-action-primary/40" />
-                        )}
-                        {holdProgress > 40 && (
-                            <div className="absolute right-6 bottom-6 h-10 w-0.5 -rotate-[30deg] bg-action-primary/40" />
-                        )}
-                        {holdProgress > 60 && (
-                            <div className="absolute bottom-4 left-8 h-6 w-0.5 rotate-12 bg-action-primary/40" />
-                        )}
-
-                        {/* Gift icon */}
-                        <div className="absolute inset-0 flex items-center justify-center">
-                            <div
-                                className={`rounded-full bg-action-primary p-3 shadow-lg transition-transform ${holdProgress > 30 ? 'animate-bounce' : ''}`}
-                            >
-                                <Icon name="gift" size={24} className="text-white" />
-                            </div>
-                        </div>
-                    </div>
-
-                    {/* Ribbon bow */}
-                    <div className="absolute -top-3 left-1/2 -translate-x-1/2">
-                        <div className="relative">
-                            {/* Left ribbon tail */}
-                            <div
-                                className="absolute top-2 left-1/2 h-4 w-2 -translate-x-[10px] bg-action-primary transition-transform"
-                                style={{
-                                    transform: `translateX(-10px) rotate(${-20 - ribbonSpread * 0.5}deg)`,
-                                    borderRadius: '0 0 2px 2px',
-                                }}
-                            />
-                            {/* Right ribbon tail */}
-                            <div
-                                className="absolute top-2 left-1/2 h-4 w-2 translate-x-[2px] bg-action-primary transition-transform"
-                                style={{
-                                    transform: `translateX(2px) rotate(${20 + ribbonSpread * 0.5}deg)`,
-                                    borderRadius: '0 0 2px 2px',
-                                }}
-                            />
-                            {/* Left loop */}
-                            <div
-                                className="absolute -top-1 -left-5 h-4 w-6 rounded-full bg-action-primary shadow-sm transition-transform"
-                                style={{ transform: `rotate(${-25 - ribbonSpread}deg)` }}
-                            />
-                            {/* Right loop */}
-                            <div
-                                className="absolute -top-1 -right-5 h-4 w-6 rounded-full bg-action-primary shadow-sm transition-transform"
-                                style={{ transform: `rotate(${25 + ribbonSpread}deg)` }}
-                            />
-                            {/* Center knot */}
-                            <div className="relative z-10 h-4 w-4 rounded-sm bg-action-primary shadow-md" />
-                        </div>
-                    </div>
-
-                    {/* Particles flying out */}
-                    {holdProgress > 30 && (
-                        <>
-                            <div className="absolute top-2 -right-4 animate-ping text-body-l [animation-duration:1s]">
-                                ✨
-                            </div>
-                            <div className="absolute bottom-4 -left-4 animate-ping text-body-l [animation-delay:0.2s] [animation-duration:1.2s]">
-                                ✨
-                            </div>
-                        </>
-                    )}
-                    {holdProgress > 60 && (
-                        <>
-                            <div className="absolute -top-2 right-2 animate-ping text-body-s [animation-delay:0.3s] [animation-duration:0.8s]">
-                                ⭐
-                            </div>
-                            <div className="absolute -bottom-2 left-2 animate-ping text-body-s [animation-delay:0.1s] [animation-duration:1s]">
-                                ⭐
-                            </div>
-                        </>
-                    )}
-                </div>
-            </div>
-
-            {/* Instructions */}
-            <p className="mt-6 text-center text-body-s text-foreground-secondary">{t('holdToUnwrap')}</p>
-        </div>
     )
 }
