+++ src/components/Home/perkClaim.consts.ts
@@ -1,0 +1,1 @@
+export const SURPRISE_CLAIM_COUNT_KEY = 'rewards_surprise_claim_count'
