+++ src/components/Home/perkClaim.types.ts
@@ -1,0 +1,1 @@
+export type ClaimPhase = 'idle' | 'holding' | 'opening' | 'revealed' | 'exiting'
