+++ src/components/Kyc/SumsubKycWrapper.tsx
@@ -1,42 +1,9 @@
 'use client'
 
-import { useEffect, useMemo, useState, useRef, useCallback } from 'react'
-import { useLocale, useTranslations } from 'next-intl'
-import Modal from '@/components/Global/Modal'
-import ActionModal from '@/components/Global/ActionModal'
-import { Icon, type IconName } from '@/components/Global/Icons/Icon'
-import { type ButtonVariant } from '@/components/0_Bruddle/Button'
-import Loading from '@/components/Global/Loading'
-import posthog from 'posthog-js'
-import { useModalsContext } from '@/context/ModalsContext'
-import { ANALYTICS_EVENTS } from '@/constants/analytics.consts'
 import { isCapacitor } from '@/utils/capacitor'
-import { toSumsubLocale } from '@/i18n/app/sumsub-locale'
-import { evaluateSumsubStatusEvent, type SumsubStatusEventPayload } from './sumsubStatusEvent.utils'
 import { SumsubNativeSdk } from './SumsubNativeSdk'
-import { SumsubSdkErrorView } from './SumsubSdkErrorView'
+import { SumsubWebSdkModal } from './SumsubWebSdkModal'
 import type { SumsubSdkProps } from './sumsubSdk.types'
-
-// todo: move to consts
-const SUMSUB_SDK_URL = 'https://static.sumsub.com/idensic/static/sns-websdk-builder.js'
-
-/**
- * How long the modal may sit open without the SDK launching before we stop
- * pretending it's loading and show the error UI. The Jul-16/17 outage spun
- * forever because nothing was watching: a never-launched SDK throws nothing,
- * requests nothing, and logs nothing. Generous enough for a slow script on a
- * bad connection; short enough that a user gets an actionable screen instead
- * of an infinite spinner.
- */
-const SDK_LAUNCH_TIMEOUT_MS = 20_000
-
-/**
- * Some SDK versions replay a resubmission as both `onApplicantResubmitted` and
- * its `idCheck.`-prefixed twin. They arrive in the same tick; a real retry takes
- * a user seconds at minimum, so this window separates the two without needing a
- * per-event identity the SDK does not give us.
- */
-const RESUBMIT_TWIN_WINDOW_MS = 1000
 
 /**
  * Every KYC entry point funnels through here, so this is the one place that
@@ -48,445 +15,4 @@ const RESUBMIT_TWIN_WINDOW_MS = 1000
 export const SumsubKycWrapper = (props: SumsubSdkProps) => {
     if (isCapacitor()) return <SumsubNativeSdk {...props} />
     return <SumsubWebSdkModal {...props} />
-}
-
-const SumsubWebSdkModal = ({
-    visible,
-    accessToken,
-    onClose,
-    onComplete,
-    onSubmitted,
-    onError,
-    onRefreshToken,
-    isMultiLevel,
-}: SumsubSdkProps) => {
-    const [sdkLoaded, setSdkLoaded] = useState(false)
-    const [sdkLoadError, setSdkLoadError] = useState(false)
-    const [isHelpModalOpen, setIsHelpModalOpen] = useState(false)
-    const [modalVariant, setModalVariant] = useState<'stop-verification' | 'trouble'>('trouble')
-    // Callback ref, NOT useRef: Modal is a headlessui <Transition>, which promotes
-    // tree state to Visible inside an effect — so the container mounts a commit
-    // AFTER `visible` flips true. (The Portal is not the culprit; it resolves its
-    // target synchronously.) A plain ref is not reactive, so the init effect below
-    // would read null on its only run and never launch the SDK — the Jul-16/17
-    // card outage. State re-runs the effect the moment the node attaches.
-    const [sdkContainer, setSdkContainer] = useState<HTMLDivElement | null>(null)
-    const sdkInstanceRef = useRef<SnsWebSdkInstance | null>(null)
-    const { setIsSupportModalOpen } = useModalsContext()
-    const t = useTranslations('kyc')
-    const tCommon = useTranslations('common')
-    const locale = useLocale()
-    const sumsubLocaleRef = useRef(toSumsubLocale(locale))
-
-    // callback refs to avoid stale closures in sdk init effect
-    const onCompleteRef = useRef(onComplete)
-    const onSubmittedRef = useRef(onSubmitted)
-    const onErrorRef = useRef(onError)
-    const onRefreshTokenRef = useRef(onRefreshToken)
-    const isMultiLevelRef = useRef(isMultiLevel)
-    // flips to true as soon as any "the user finished something" event fires.
-    // Drives the close-confirmation short-circuit: if the user has already
-    // submitted, tapping X closes the modal without asking "stop verification?"
-    const hasSubmittedRef = useRef(false)
-    const lastResubmitSignalRef = useRef(0)
-    // Watchdog bookkeeping: did sdk.launch() ever run for this open?
-    const hasLaunchedRef = useRef(false)
-
-    useEffect(() => {
-        onCompleteRef.current = onComplete
-        onSubmittedRef.current = onSubmitted
-        onErrorRef.current = onError
-        onRefreshTokenRef.current = onRefreshToken
-        isMultiLevelRef.current = isMultiLevel
-    }, [onComplete, onSubmitted, onError, onRefreshToken, isMultiLevel])
-
-    useEffect(() => {
-        sumsubLocaleRef.current = toSumsubLocale(locale)
-    }, [locale])
-
-    // stable wrappers that read from refs
-    const stableOnComplete = useCallback(() => onCompleteRef.current(), [])
-    const stableOnSubmitted = useCallback(() => onSubmittedRef.current?.(), [])
-    const stableOnError = useCallback((error: unknown) => onErrorRef.current?.(error), [])
-    const stableOnRefreshToken = useCallback(() => onRefreshTokenRef.current(), [])
-
-    // load sumsub websdk script
-    useEffect(() => {
-        if (window.snsWebSdk) {
-            setSdkLoaded(true)
-            return undefined
-        }
-
-        const handleLoaded = () => setSdkLoaded(true)
-        const handleError = () => {
-            console.error(new Error('[sumsub] failed to load websdk script'))
-            setSdkLoadError(true)
-        }
-
-        const existingScript = document.getElementById('sumsub-websdk')
-        if (existingScript) {
-            // another wrapper instance appended the script and it's still
-            // downloading — a bare existence check would init against an
-            // undefined window.snsWebSdk
-            existingScript.addEventListener('load', handleLoaded)
-            existingScript.addEventListener('error', handleError)
-            // the script may have finished between the snsWebSdk check above
-            // and the listener attach — re-check so we don't wait forever
-            if (window.snsWebSdk) handleLoaded()
-            return () => {
-                existingScript.removeEventListener('load', handleLoaded)
-                existingScript.removeEventListener('error', handleError)
-            }
-        }
-
-        const script = document.createElement('script')
-        script.id = 'sumsub-websdk'
-        script.src = SUMSUB_SDK_URL
-        script.async = true
-        script.onload = handleLoaded
-        script.onerror = handleError
-        document.head.appendChild(script)
-        return undefined
-    }, [])
-
-    // initialize sdk as soon as the modal is visible and all deps are ready
-    useEffect(() => {
-        if (!visible || !accessToken || !sdkLoaded || !sdkContainer) return
-
-        // clean up previous instance
-        if (sdkInstanceRef.current) {
-            try {
-                sdkInstanceRef.current.destroy()
-            } catch {
-                // ignore cleanup errors
-            }
-        }
-
-        // declared outside try so cleanup can access it
-        let iframeObserver: MutationObserver | null = null
-
-        try {
-            // track sdk init time so we can ignore stale onApplicantStatusChanged events
-            // that fire immediately when the applicant is already approved (e.g. additional-docs flow)
-            const sdkInitTime = Date.now()
-
-            const handleSubmitted = () => {
-                console.log('[sumsub] onApplicantSubmitted fired')
-                const isFirstSubmit = !hasSubmittedRef.current
-                hasSubmittedRef.current = true
-                // for multi-level workflows (LATAM/EU), the SDK transitions to
-                // Level 2 internally. don't close the modal on Level 1 submission —
-                // but do report it, or the session never emits a submit signal
-                // (onComplete only fires when a close is wanted, and the APPROVED
-                // close skips it).
-                if (isMultiLevelRef.current) {
-                    if (isFirstSubmit) stableOnSubmitted()
-                    return
-                }
-                stableOnComplete()
-            }
-            // resubmission = user retried after rejection (ACTION_REQUIRED).
-            // Multi-level keeps the SDK open: a retry on Level 1 still owes the
-            // follow-up questionnaire, so closing here stranded the applicant
-            // looking submitted with a level outstanding.
-            //
-            // EVERY logical retry reports, unlike handleSubmitted's first-only
-            // gate. Downstream this is the attempt boundary that lets a repeated
-            // rejection be reported again, so suppressing it after the first
-            // submission swallowed the retry's own REJECTED. Only the SDK's
-            // duplicate twin event is collapsed.
-            const handleResubmitted = () => {
-                console.log('[sumsub] onApplicantResubmitted fired')
-                hasSubmittedRef.current = true
-                if (isMultiLevelRef.current) {
-                    const now = Date.now()
-                    if (now - lastResubmitSignalRef.current >= RESUBMIT_TWIN_WINDOW_MS) {
-                        lastResubmitSignalRef.current = now
-                        stableOnSubmitted()
-                    }
-                    return
-                }
-                stableOnComplete()
-            }
-            // Applicant Actions (like rain-card-application) emit this instead
-            // of onApplicantSubmitted. Without the listener we'd miss the
-            // signal and the close button would keep warning about
-            // abandonment after a successful action submission.
-            const handleActionSubmitted = () => {
-                console.log('[sumsub] action submitted fired')
-                const isFirstSubmit = !hasSubmittedRef.current
-                hasSubmittedRef.current = true
-                if (isMultiLevelRef.current) {
-                    if (isFirstSubmit) stableOnSubmitted()
-                    return
-                }
-                stableOnComplete()
-            }
-            // RED stays open so the user can resubmit; the resubmission path
-            // emits onApplicantActionSubmitted, which handleActionSubmitted
-            // closes. See evaluateSumsubStatusEvent for the early-guard rules.
-            const handleStatusEvent = (eventName: string) => (payload: SumsubStatusEventPayload) => {
-                console.log(`[sumsub] ${eventName} fired`, payload)
-                const evaluation = evaluateSumsubStatusEvent({
-                    payload,
-                    sdkInitTime,
-                    now: Date.now(),
-                    isMultiLevel: !!isMultiLevelRef.current,
-                })
-                if (evaluation.markSubmitted) hasSubmittedRef.current = true
-                if (evaluation.autoClose) stableOnComplete()
-            }
-            const handleStatusChanged = handleStatusEvent('onApplicantStatusChanged')
-            const handleActionCompleted = handleStatusEvent('onApplicantActionStatusChanged')
-
-            const sdk = window.snsWebSdk
-                .init(accessToken, stableOnRefreshToken)
-                .withConf({ lang: sumsubLocaleRef.current, theme: 'light' })
-                .withOptions({ addViewportTag: false, adaptIframeHeight: true })
-                .on('onApplicantSubmitted', handleSubmitted)
-                .on('onApplicantResubmitted', handleResubmitted)
-                .on('onApplicantStatusChanged', handleStatusChanged)
-                // Applicant Action events (card-application, additional-docs, etc.)
-                .on('onActionSubmitted', handleActionSubmitted)
-                .on('onApplicantActionSubmitted', handleActionSubmitted)
-                .on('onApplicantActionStatusChanged', handleActionCompleted)
-                // also listen for idCheck-prefixed events (some sdk versions use these)
-                .on('idCheck.onApplicantSubmitted', handleSubmitted)
-                .on('idCheck.onApplicantResubmitted', handleResubmitted)
-                .on('idCheck.onApplicantStatusChanged', handleStatusChanged)
-                .on('idCheck.onActionSubmitted', handleActionSubmitted)
-                .on('idCheck.onApplicantActionSubmitted', handleActionSubmitted)
-                .on('idCheck.onApplicantActionStatusChanged', handleActionCompleted)
-                .on('onError', (error: unknown) => {
-                    console.error('[sumsub] sdk error', error)
-                    stableOnError(error)
-                })
-                .build()
-
-            sdk.launch(sdkContainer)
-            sdkInstanceRef.current = sdk
-            // The positive signal. Without it, "opened" is the only thing we
-            // record and a total launch failure is indistinguishable from
-            // nobody bothering to verify.
-            hasLaunchedRef.current = true
-            posthog.capture(ANALYTICS_EVENTS.KYC_SDK_LAUNCHED, { isMultiLevel: !!isMultiLevelRef.current })
-
-            // ensure the sdk-created iframe gets camera/microphone permissions.
-            // some sdk versions don't set the allow attribute, which blocks
-            // media device access in cross-origin iframes.
-            iframeObserver = new MutationObserver((mutations) => {
-                for (const mutation of mutations) {
-                    for (const node of mutation.addedNodes) {
-                        if (node instanceof HTMLIFrameElement && !node.allow?.includes('camera')) {
-                            node.allow = 'camera; microphone; fullscreen'
-                        }
-                    }
-                }
-            })
-            iframeObserver.observe(sdkContainer, { childList: true })
-
-            // also patch any iframe that was added before the observer
-            const existingIframe = sdkContainer.querySelector('iframe')
-            if (existingIframe && !existingIframe.allow?.includes('camera')) {
-                existingIframe.allow = 'camera; microphone; fullscreen'
-            }
-        } catch (error) {
-            console.error('[sumsub] failed to initialize sdk', error)
-            // surface the error UI — without this the modal stays blank
-            setSdkLoadError(true)
-            posthog.capture(ANALYTICS_EVENTS.KYC_SDK_INIT_FAILED, {
-                message: error instanceof Error ? error.message : String(error),
-            })
-            stableOnError(error)
-        }
-
-        return () => {
-            iframeObserver?.disconnect()
-            if (sdkInstanceRef.current) {
-                try {
-                    sdkInstanceRef.current.destroy()
-                } catch {
-                    // ignore cleanup errors
-                }
-                sdkInstanceRef.current = null
-            }
-        }
-    }, [
-        visible,
-        accessToken,
-        sdkLoaded,
-        sdkContainer,
-        stableOnComplete,
-        stableOnSubmitted,
-        stableOnError,
-        stableOnRefreshToken,
-    ])
-
-    // reset state when modal closes (the init effect's cleanup already
-    // destroys the SDK instance — visible is one of its deps)
-    useEffect(() => {
-        if (!visible) {
-            setSdkLoadError(false)
-            hasSubmittedRef.current = false
-            hasLaunchedRef.current = false
-        }
-    }, [visible])
-
-    // Watchdog: the modal is open but the SDK never launched.
-    //
-    // Deliberately keyed on `visible` ALONE, not on the init effect's deps. Every
-    // way this component can fail silently — a null container, a token that never
-    // arrives, a script that never loads — ends with the init effect simply not
-    // running, so anything that watches those deps would also never run. Watching
-    // the one fact the user cares about ("I opened it and nothing happened")
-    // catches the whole class, including failures we haven't thought of yet.
-    useEffect(() => {
-        if (!visible) return
-        const timer = setTimeout(() => {
-            if (hasLaunchedRef.current) return
-            console.error('[sumsub] sdk never launched within timeout — surfacing error')
-            posthog.capture(ANALYTICS_EVENTS.KYC_SDK_LAUNCH_TIMEOUT, {
-                timeoutMs: SDK_LAUNCH_TIMEOUT_MS,
-                // The three deps whose absence explains every silent stall.
-                hadAccessToken: !!accessToken,
-                sdkScriptLoaded: sdkLoaded,
-                hadContainer: !!sdkContainer,
-            })
-            setSdkLoadError(true)
-        }, SDK_LAUNCH_TIMEOUT_MS)
-        return () => clearTimeout(timer)
-        // accessToken/sdkLoaded/sdkContainer are read for diagnostics only; they
-        // must not restart the timer or a late-arriving dep would reset the clock.
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [visible])
-
-    // Close-button handler. After the user has submitted, the "are you sure
-    // you want to stop?" modal is misleading — they're done, not abandoning.
-    // Skip straight to onClose in that case.
-    const handleCloseButton = useCallback(() => {
-        if (hasSubmittedRef.current) {
-            onClose()
-            return
-        }
-        setModalVariant('stop-verification')
-        setIsHelpModalOpen(true)
-    }, [onClose])
-
-    const modalDetails = useMemo(() => {
-        if (modalVariant === 'trouble') {
-            return {
-                title: t('wrapper.troubleTitle'),
-                description: t('wrapper.troubleDescription'),
-                icon: 'question-mark' as IconName,
-                iconContainerClassName: 'bg-action-primary',
-                ctas: [
-                    {
-                        text: t('wrapper.chatWithSupport'),
-                        onClick: () => setIsSupportModalOpen(true),
-                        variant: 'purple' as ButtonVariant,
-                        shadowSize: '4' as const,
-                    },
-                    {
-                        text: tCommon('cancel'),
-                        onClick: () => setIsHelpModalOpen(false),
-                        variant: 'stroke' as ButtonVariant,
-                        className: 'w-full',
-                    },
-                ],
-            }
-        }
-
-        return {
-            title: t('wrapper.exitForNowTitle'),
-            description: t('wrapper.exitForNowDescription'),
-            icon: 'alert' as IconName,
-            iconContainerClassName: 'bg-action-secondary',
-            ctas: [
-                {
-                    text: t('wrapper.exit'),
-                    onClick: () => {
-                        setIsHelpModalOpen(false)
-                        onClose()
-                    },
-                    variant: 'purple' as ButtonVariant,
-                    shadowSize: '4' as const,
-                },
-                {
-                    text: tCommon('continue'),
-                    onClick: () => setIsHelpModalOpen(false),
-                    variant: 'stroke' as ButtonVariant,
-                    className: 'w-full',
-                },
-            ],
-        }
-    }, [modalVariant, onClose, setIsSupportModalOpen, t, tCommon])
-
-    return (
-        <>
-            <Modal
-                visible={visible}
-                onClose={onClose}
-                classWrap="h-full w-full !max-w-none sm:!max-w-[600px] border-none sm:m-auto m-0"
-                classOverlay={`bg-black/50 ${isHelpModalOpen ? 'pointer-events-none' : ''}`}
-                video={false}
-                className={`z-[100] !p-0 md:!p-6 ${isHelpModalOpen ? 'pointer-events-none' : ''}`}
-                classButtonClose="hidden"
-                preventClose={true}
-                hideOverlay={false}
-            >
-                {sdkLoadError ? (
-                    <SumsubSdkErrorView onClose={onClose} message={t('wrapper.loadError')} />
-                ) : (
-                    <div className="flex h-full flex-col">
-                        <div className="flex items-center justify-between px-4 py-2">
-                            <button
-                                type="button"
-                                aria-label={tCommon('contactSupport')}
-                                onClick={() => {
-                                    setModalVariant('trouble')
-                                    setIsHelpModalOpen(true)
-                                }}
-                                className="relative flex items-center gap-1 p-1 transition-opacity duration-instant after:absolute after:-inset-2 focus-visible:outline-[3px] focus-visible:outline-action-focus active:opacity-60"
-                            >
-                                <Icon name="peanut-support" size={20} className="text-foreground-secondary" />
-                            </button>
-                            <button
-                                type="button"
-                                aria-label={tCommon('close')}
-                                onClick={handleCloseButton}
-                                className="relative p-1 transition-opacity duration-instant after:absolute after:-inset-2 focus-visible:outline-[3px] focus-visible:outline-action-focus active:opacity-60"
-                            >
-                                <Icon name="cancel" size={24} />
-                            </button>
-                        </div>
-                        <div className="relative w-full flex-1">
-                            {/* sits behind the SDK iframe — covered once it paints */}
-                            <div className="absolute inset-0 flex items-center justify-center">
-                                <Loading className="h-8 w-8" />
-                            </div>
-                            <div
-                                ref={setSdkContainer}
-                                className="relative h-full w-full overflow-auto [&>iframe]:!min-h-full"
-                            />
-                        </div>
-                    </div>
-                )}
-            </Modal>
-            {/* rendered outside the outer Modal to avoid pointer-events-none blocking clicks */}
-            <ActionModal
-                visible={isHelpModalOpen}
-                onClose={() => setIsHelpModalOpen(false)}
-                title={modalDetails.title}
-                description={modalDetails.description}
-                icon={modalDetails.icon}
-                iconContainerClassName={modalDetails.iconContainerClassName}
-                modalPanelClassName="max-w-full"
-                ctaClassName="grid grid-cols-1 gap-3"
-                contentContainerClassName="px-6 py-6"
-                modalClassName="!z-[10001]"
-                preventClose={true}
-                ctas={modalDetails.ctas}
-            />
-        </>
-    )
 }
