+++ src/features/add-money/types.ts
@@ -1,0 +1,2 @@
+// Step type for URL state
+export type BridgeBankStep = 'verify' | 'inputAmount' | 'showDetails'
